# sdks/sdkwork-messaging-backend-sdk/openapi/sdkwork-messaging-backend-api.openapi.yaml
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Messaging Backend API",
    "version": "1.0.0",
    "description": "Backend/admin contract for notification creation, announcement publishing, app push delivery, SMS/email sends, verification policies, channel configuration, and message templates.",
    "x-sdkwork-api-authority": "sdkwork-messaging-backend-api",
    "x-sdkwork-sdk-family": "sdkwork-messaging-backend-sdk"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local sdkwork-messaging runtime"
    }
  ],
...
  "security": [
    {
      "AuthToken": [],
      "AccessToken": []
    }
  ],
  "paths": {
...
    "/backend/v3/api/messaging/verification_policies": {
      "get": {
        "tags": [
          "messaging"
        ],
        "summary": "messaging.verificationPolicies.list",
        "operationId": "messaging.verificationPolicies.list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "additionalProperties": false,
                          "required": [
                            "items",
                            "pageInfo"
                          ],
                          "properties": {
                            "items": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/MessagingVerificationPolicy"
                              }
                            },
                            "pageInfo": {
                              "$ref": "#/components/schemas/PageInfo"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-messaging",
        "x-sdkwork-api-authority": "sdkwork-messaging-backend-api",
        "x-sdkwork-domain": "messaging",
        "x-sdkwork-resource": "messaging.verificationPolicies",
        "x-sdkwork-permission": "messaging.verificationPolicies.get",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-data-scope": "organization",
        "x-sdkwork-audit-event": "messaging.verification_policies.list",
        "x-sdkwork-idempotent": false,
        "x-sdkwork-deployment": "all",
        "x-sdkwork-source": "crates/sdkwork-routes-messaging-backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-routes-messaging-backend-api",
        "x-sdkwork-server-request-id": true
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token"
      }
    },
    "schemas": {
...
      "MessagingVerificationPolicy": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "sceneCode",
          "channel",
          "ttlSeconds",
          "maxAttempts",
          "enabled"
        ],
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "sceneCode": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "channel": {
            "type": "string",
            "enum": [
              "sms",
              "email"
            ]
          },
          "ttlSeconds": {
            "type": "integer",
            "minimum": 30
          },
          "maxAttempts": {
            "type": "integer",
            "minimum": 1,
            "maximum": 20
          },
          "messageSubject": {
            "type": "string"
          },
          "messageBodyPattern": {
            "type": "string",
            "minLength": 1
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
...
      "PageInfo": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "mode"
        ],
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "offset",
              "cursor"
            ]
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200
          },
          "totalItems": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/SdkWorkPlatformErrorCode"
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "i18nKey": {
            "type": "string",
            "description": "Optional stable localization key such as errors.result.40001."
          },
          "locale": {
            "type": "string",
            "description": "Optional effective BCP 47 locale used by framework message mapping."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
      "SdkWorkApiResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "enum": [
              0
            ],
            "default": 0,
            "minimum": 0,
            "maximum": 0
          },
          "data": {
            "description": "Operation-specific payload typed per response schema."
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          }
        }
      },
...
      "SdkWorkPlatformErrorCode": {
        "type": "integer",
        "format": "int32",
        "minimum": 40001,
        "maximum": 79999,
        "description": "Platform or domain error code per API_SPEC.md §15.3."
      },
      "FieldError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "int32",
            "minimum": 40011,
            "maximum": 40099
          },
          "i18nKey": {
            "type": "string"
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                }
              ]
            }
          }
        }
      },
...
    }
  }
}